// frontend/src/api/models/memberDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */

export interface MemberDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  originalName?: string | null;
  displayOrder?: number;
  /** @nullable */
  linkedUserId?: string | null;
  /** @nullable */
  linkedUserDisplayName?: string | null;
  /** @nullable */
  linkedUserAvatarUrl?: string | null;
  /** @nullable */
  claimedAt?: string | null;
}