// frontend/src/api/models/expenseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { ExpenseItemDto } from "./expenseItemDto";

export interface ExpenseDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  amount?: number;
  serviceFeePercent?: number;
  isItemized?: boolean;
  /** @nullable */
  paidById?: string | null;
  /** @nullable */
  participantIds?: string[] | null;
  /** @nullable */
  items?: ExpenseItemDto[] | null;
  createdAt?: string;
}