// frontend/src/api/models/expenseItemDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */

export interface ExpenseItemDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  amount?: number;
  /** @nullable */
  paidById?: string | null;
  /** @nullable */
  participantIds?: string[] | null;
}