// frontend/src/api/models/conflictInfo.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { ConflictInfoLocalValue } from "./conflictInfoLocalValue";
import type { ConflictInfoServerValue } from "./conflictInfoServerValue";
import type { ConflictInfoResolvedValue } from "./conflictInfoResolvedValue";

export interface ConflictInfo {
  /** @nullable */
  type?: string | null;
  /** @nullable */
  entityId?: string | null;
  /** @nullable */
  field?: string | null;
  /** @nullable */
  localValue?: ConflictInfoLocalValue;
  /** @nullable */
  serverValue?: ConflictInfoServerValue;
  /** @nullable */
  resolution?: string | null;
  /** @nullable */
  resolvedValue?: ConflictInfoResolvedValue;
}