// frontend/src/api/models/deltaIdMappingsDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { DeltaIdMappingsDtoMembers } from "./deltaIdMappingsDtoMembers";
import type { DeltaIdMappingsDtoExpenses } from "./deltaIdMappingsDtoExpenses";
import type { DeltaIdMappingsDtoExpenseItems } from "./deltaIdMappingsDtoExpenseItems";

export interface DeltaIdMappingsDto {
  /** @nullable */
  members?: DeltaIdMappingsDtoMembers;
  /** @nullable */
  expenses?: DeltaIdMappingsDtoExpenses;
  /** @nullable */
  expenseItems?: DeltaIdMappingsDtoExpenseItems;
}